+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add the `$merge` parameter on `Scope::setUser` to allow merging user context. (#929)
 - Make the `integrations` option accept a `callable` that will receive the list of default integrations and returns a customized list (#919)
 - Add the `IgnoreErrorsIntegration` integration to deprecate and replace the `exclude_exceptions` option (#928)
+- Allow setting custom contexts on the scope and on the event (#839)
 
 ## 2.2.6 (2019-12-18)